FNF Cutscene Lua script
ok, It's my first Lua scripts that I'm upload to Gamebanana. I think that cutscenes for FNF mods is one of the main parts of the mod and I think that modders know how to write it. But I think that some people can't write a code, so I'll write my own code that you can download and use in your mod. And here's what you need to do:  1) You must upload your video into "videos" folder in Psych Engine.  2) Your video must be size 1280x720.  Have fun !

P.s. The problem of this code only in message: "return value not supported" that I've got when I play song where I have cutscene, but this do nothing for game and songs, so I think that this isn't problem and you can ignore it